Changing datatype for the isbn column

+++ app/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Date, String, Integer, Double, func
+from sqlalchemy import Column, Date, String, Integer, Double, BigInteger
 from sqlalchemy.orm import declarative_base
 
 Base = declarative_base()
@@ -7,7 +7,7 @@ metadata = Base.metadata
 class Books(Base):
     __tablename__ = "books"
 
-    isbn = Column(Integer, primary_key=True)
+    isbn = Column(BigInteger, primary_key=True)
     title = Column(String(60), unique=True)
     authors = Column(String(60))
     average_rating = Column(Double)
